pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eagle-acquisition-plugin"
version = "2.0.0"
description = "NCI Office of Acquisitions plugin — 25 skills, 6 agents, full acquisition lifecycle"
requires-python = ">=3.11"
license = {text = "CC0-1.0"}

dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "boto3>=1.34.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
web = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.13.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
